### 2024-03-28 #6550 @editorialbot
[PRE REVIEW]: Satisfiability.jl: Satisfiability Modulo Theories in Julia
**Submitting author:** @elsoroka (<a href="http://orcid.org/0009-0001-2710-469X">Emiko Soroka</a>)
**Repository:** https://github.com/elsoroka/Satisfiability.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.1
