### 2024-03-28 #6550 @danielskatz
👋 @elsoroka - thanks for your submission. I'll next work on finding an editor.

While I do so, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Also please fix the invalid DOIs. Make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
