+++ joss-paper/bibliography.bib
@@ -1,12 +1,12 @@
-\@article{convexjl,
- title = {Convex Optimization in {J}ulia},
- author = {Udell, Madeleine and Mohan, Karanveer and Zeng, David and Hong, Jenny and Diamond, Steven and Boyd, Stephen},
- year = {2014},
- journal = {SC14 Workshop on High Performance Technical Computing in Dynamic Languages},
- archivePrefix = "arXiv",
- eprint = {1410.4821},
- primaryClass = "math-oc",
- doi={10.48550/arXiv.1410.4821},
+@article{convexjl,
+ 	title = {Convex Optimization in {J}ulia},
+	author = {Udell, Madeleine and Mohan, Karanveer and Zeng, David and Hong, Jenny and Diamond, Steven and Boyd, Stephen},
+	year = {2014},
+	journal = {SC14 Workshop on High Performance Technical Computing in Dynamic Languages},
+	archivePrefix = "arXiv",
+	eprint = {1410.4821},
+	primaryClass = "math-oc",
+	doi={10.48550/arXiv.1410.4821},
 }
 
 @article{Lubin2023,
@@ -95,8 +95,8 @@
 	pages     = {415--442},
 	publisher = {Springer},
 	year      = {2022},
-	url       = {https://doi.org/10.1007/978-3-030-99524-9\_24},
-	doi       = {10.1007/978-3-030-99524-9\_24},
+	url       = {https://doi.org/10.1007/978-3-030-99524-9_24},
+	doi       = {10.1007/978-3-030-99524-9_24},
 	timestamp = {Fri, 01 Apr 2022 15:49:27 +0200},
 	biburl    = {https://dblp.org/rec/conf/tacas/BarbosaBBKLMMMN22.bib},
 	bibsource = {dblp computer science bibliography, https://dblp.org},
@@ -120,10 +120,17 @@
 	booktitle={Proceedings of the 8th ACM SIGPLAN International Symposium on Scala},
 	pages={51--55},
 	year={2017},
-	doi={3136000.3136004}
+	doi={10.1145/3136000.3136004}
 }
 
-@misc{picosat_jl, url={https://github.com/sisl/PicoSAT.jl}, author={Bolewski, Jake and Lucibello, Carlo and Bouton, Maxime}, publisher={Stanford Intelligent Systems Laboratory}, year={2020}, month={Jul}} 
+@misc{picosat_jl,
+	title={PicoSAT.jl},
+	url={https://github.com/sisl/PicoSAT.jl},
+	author={Bolewski, Jake and Lucibello, Carlo and Bouton, Maxime},
+	publisher={Stanford Intelligent Systems Laboratory},
+	year={2020},
+	month={Jul}
+} 
 
 @TECHREPORT{smtlib2,
 	author =	 {Clark Barrett and Pascal Fontaine and Cesare Tinelli},
@@ -139,7 +146,8 @@
 	booktitle={International Conference on Verification, Model Checking, and Abstract Interpretation},
 	pages={91--105},
 	year={2023},
-	organization={Springer}
+	organization={Springer},
+	doi={10.1007/978-3-031-24950-1_5}
 }
 
 @inproceedings{saouli2023cosysel,
@@ -148,7 +156,8 @@
 	booktitle={International Conference on Verification, Model Checking, and Abstract Interpretation},
 	pages={252--266},
 	year={2023},
-	organization={Springer}
+	organization={Springer},
+	doi={10.1007/978-3-031-24950-1_12}
 }
 
 @article{bezanson2017julia,
@@ -160,54 +169,7 @@
 	pages={65--98},
 	year={2017},
 	publisher={SIAM},
-	url={https://doi.org/10.1137/141000671}
-}
-
[…]
