### 2024-04-12 #6550 @diehlpk
> Hi @diehlpk , I do have time to review, but I don't know if I'm the right person! This is the first time I'm hearing the words "satisfiability modulo theories", so my input might not be so helpful...

I think your input one the code quality, installation process, and documentation would be very helpful.
