### 2024-04-14 #6550 @elsoroka
@diehlpk, @mykelk suggested this list of people who have contributed to SMT projects in GitHub as potential reviewers.

https://github.com/marcogario
https://github.com/mikand
https://github.com/NikolajBjorner
https://github.com/wintersteiger
https://github.com/ajreynol
https://github.com/aniemetz
