### 2024-04-22 9fdc044
An example I thought of while working on another project. We use Z3 to convert a formula to CNF and some metaprogramming tricks to get it back into its Satisfiability.jl form.
