### 2024-05-23 2ca0b80
@Mykel suggested this change to automatically download z3 if it's not already installed.
