### 2024-06-07 elsoroka/Satisfiability.jl#49 @rafaelbailo
JOSS Review Comments
Hello, and congrats on your JOSS submission! Overall, the package and the paper are in a very good state, and I've ticked most of the items on my [reviewer checklist](https://github.com/openjournals/joss-reviews/issues/6757?notification_referrer_id=NT_kwDOATudUrQxMDY4OTgxNTAzNjoyMDY4NDExNA#issuecomment-2115380330) already.

I have a few minor comments that shouldn't take too long to address. Please let me know if these make sense, I'm happy to discuss them further if needed:

## The Paper

* [x] 1. No code examples are given in the paper. Might it be nice to include some of the examples from the `README.md` file? (See comments below, feel free to ignore this point or just include a link to the examples in the paper, to avoid code possibly becoming outdated in the paper).
* [x] 2. I found it strange to have the "The SMT-LIB specification language" featured so prominently in the paper, only to conclude by stating that knowledge of SMT-LIB is not needed to use the package. It might be worth demoting this section to another part of the paper, in order to focus on your own work.

## The Repository

* [x] 1. The first example is broken do to a wrong quotation character. This is fixed [here](https://github.com/elsoroka/Satisfiability.jl/pull/47).
* [x] 2. The third example is broken, see the [issue](https://github.com/elsoroka/Satisfiability.jl/issues/48).
* [x] 3. Please create a `CONTRIBUTING.md` file with contribution guidelines, which will be useful for those contributing to your open issues.
* [x] 4. Similarly, please consider using [github issue templates](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) for bug reporting.

## The Code

* [x] 1. Running the tests produces some Error and Warning output, even though all the tests pass. Is this expected?
* [x] 2. There are plenty of examples in the `examples` folder, but these are not tested. You might want to add unit tests that run each example and check that no warnings are produced, to ensure that they don't become outdated inadvertently.
* [x] 3. You have great test coverage already. I see that most of the lines missing coverage are the if statements for error handling; Julia does provide unit testing for errors through the [`@test_throws`](https://docs.julialang.org/en/v1/stdlib/Test/#Basic-Unit-Tests) macro. It wouldn't take long to test that the error handling behaves as expected, and it would bring your coverage way up!

## The Documentation

* [x] 1. It seems that a recent update means no solvers have to be installed before using your package. However, the documentation still features installation instructions for solvers. Might this be clarified?
* [x] 2. There are even more examples in the documentation, which is helpful, but as far as I can see these are not tested either. You might want to use [Doctests](https://documenter.juliadocs.org/stable/man/doctests/).
* [x] 3. There are documentation examples, `examples` folder examples, and `paper_examples` examples, all which appear different. Would it make sense to maintain the same examples in the repository and the documentation? Also, since the JOSS paper currently features no examples, I assume `paper_examples` are destined for another publication? If that's the case, it might make sense to place them in a different repository.
