### 2024-06-07 #6757 @rafaelbailo
Hi @diehlpk, I've just posted my review in a [separate issue](https://github.com/elsoroka/Satisfiability.jl/issues/49). Thanks for your patience!
