### 2024-06-07 elsoroka/Satisfiability.jl#49 @diehlpk
@elsoroka please have a look at these comments.
