### 2024-06-07 elsoroka/Satisfiability.jl#49 @rafaelbailo
> > * [ ]  1. No code examples are given in the paper. Might it be nice to include some of the examples from the `README.md` file?
>
> I do not recommend adding code examples to the paper since these might not work with future versions. Maybe a link to the examples in the repo is sufficient?

That's fair enough!
