### 2024-06-11 d67e74e
added Mykel Kochenderfer's function to convert unicode Julia variables to ASCII variables compatible with SMT-LIB.
