+++ src/smt_representation.jl
@@ -1,3 +1,5 @@
+using Unicode
+
 # Mapping of Julia Expr types to SMT names. This is necessary because to distinguish from native types Bool, Int, Real, etc, we call ours BoolExpr, IntExpr, RealExpr, etc.
 __smt_typestr(e::BoolExpr) = "Bool"
 __smt_typestr(e::IntExpr) = "Int"
@@ -43,7 +45,7 @@ __smt_generated_ops = Dict(
     :sign_extend  => (e::SlicedBitVectorExpr)   -> "(_ sign_extend $(e.range))",
     :rotate_left  => (e::SlicedBitVectorExpr)   -> "(_ rotate_left $(e.range))",
     :rotate_right => (e::SlicedBitVectorExpr)   -> "(_ rotate_right $(e.range))",
-    :ufunc        => (e::AbstractExpr) -> split(e.name, "_")[1]
+    :ufunc        => (e::AbstractExpr) -> split(convert_to_ascii(e.name), "_")[1]
 )
 
 # Finally, we provide facilities for correct encoding of consts
@@ -62,6 +64,40 @@ function __format_smt_const(exprtype::Type, c::AbstractExpr)
     end
 end
 
+# Conversion of unicode Julia variable names to ASCII and back
+function convert_to_ascii(input_string::String)::String
+    output_string = ""
+    for char in input_string
+        if char <= '\x7F'
+            output_string *= char
+        else
+            # Use a placeholder for non-ASCII characters
+            unicode_repr = string(Char(char))
+            placeholder = string("|u+", hex(Char(char), 4), "|")
+            output_string *= placeholder
+        end
+    end
+    return output_string
+end
+
+function decode_from_ascii(ascii_string::String)::String
+    output_string = ""
+    i = 1
+    while i <= length(ascii_string)
+        if i <= length(ascii_string) - 8 && ascii_string[i:i+2] == "|u+" && ascii_string[i+7] == '|'
+            # Extract Unicode code point from placeholder
+            code_point_str = ascii_string[i+3:i+6]
+            code_point = parse(Int, code_point_str, base=16)
+            output_string *= Char(code_point)
+            i += 8  # Skip past the placeholder
+        else
+            output_string *= ascii_string[i]
+            i += 1
+        end
+    end
+    return output_string
+end
+
 
 ##### GENERATING SMTLIB REPRESENTATION #####
 
@@ -75,13 +111,13 @@ Examples:
 * `declare(and(z1, z2))` returns `"(declare-fun z1 () Bool)\\n(declare-fun z2 () Bool)\\n"`.
 =#
 function declare(z::AbstractExpr; line_ending='\n')
-    return "(declare-fun $(z.name) () $(__smt_typestr(z)))"
+    return "(declare-fun $(convert_to_ascii(z.name)) () $(__smt_typestr(z)))"
 end
 
 declare(zs::Array{T}) where T <: AbstractExpr = reduce(*, declare.(zs; line_ending=line_ending))
 
 function declare_ufunc(e::AbstractExpr)
-        name = e.children[1].name
+        name = convert_to_ascii(e.children[1].name)
         intype = __smt_typestr(e.children[2])
         outtype = __smt_typestr(e)
         return "(declare-fun $name($intype) $outtype)"
@@ -97,9 +133,9 @@ function __get_smt_name(z::AbstractExpr)
     global GLOBAL_VARNAMES
     appears_in = map( (t) -> z.name ∈ GLOBAL_VARNAMES[t], __EXPR_TYPES)
     if sum(appears_in) > 1
-        return "(as $(z.name) $(__smt_typestr(z)))"
+        return "(as $(convert_to_ascii(z.name)) $(__smt_typestr(z)))"
     else # easy case, one variable with z.name is defined
-        return z.name
[…]
