### 2024-06-11 9878c13
fixed chatGPT's unicode conversion function and added tests for it
