+++ test/int_real_tests.jl
@@ -48,51 +48,51 @@ using Test
 end
 
 @testset "Construct n-ary ops" begin
-    @satvariable(a, Int)
+    @satvariable(α, Int)
     @satvariable(b[1:2], Int)
-    @satvariable(ar, Real)
+    @satvariable(αr, Real)
     @satvariable(br[1:2], Real)
 
     # Operations with expressions only
-    @test all(isa.(a .+ b, IntExpr))
-    @test all(isa.(b .- a, IntExpr))
-    @test all(isa.(br .* ar, RealExpr))
-    @test all(isa.(a ./ b, RealExpr))
+    @test all(isa.(α .+ b, IntExpr))
+    @test all(isa.(b .- α, IntExpr))
+    @test all(isa.(br .* αr, RealExpr))
+    @test all(isa.(α ./ b, RealExpr))
 
-    @test isequal(a*a*a, a^3)
-    @test isequal(a^(-1), 1.0/to_real(a))
-    @test isequal((1.0/ar)*(1.0/ar), ar^(-2))
+    @test isequal(α*α*α, α^3)
+    @test isequal(α^(-1), 1.0/to_real(α))
+    @test isequal((1.0/αr)*(1.0/αr), αr^(-2))
 
     # Operations with mixed constants and type promotion
     # Adding Int and Bool types results in an IntExpr
-    children = [a, IntExpr(:const, AbstractExpr[], 2, "const_2")]
-    @test isequal(sum([a, 1, true]), IntExpr(:add, children, nothing, Satisfiability.__get_hash_name(:add, children, is_commutative=true)))
+    children = [α, IntExpr(:const, AbstractExpr[], 2, "const_2")]
+    @test isequal(sum([α, 1, true]), IntExpr(:add, children, nothing, Satisfiability.__get_hash_name(:add, children, is_commutative=true)))
     
     # Type promotion to RealExpr works when we add a float-valued literal
-    children = [a, RealExpr(:const, AbstractExpr[], 3., "const_3.0")]
-    @test isequal(sum([1.0, a, true, 1]), RealExpr(:add, children, nothing, Satisfiability.__get_hash_name(:add, children, is_commutative=true)))
+    children = [α, RealExpr(:const, AbstractExpr[], 3., "const_3.0")]
+    @test isequal(sum([1.0, α, true, 1]), RealExpr(:add, children, nothing, Satisfiability.__get_hash_name(:add, children, is_commutative=true)))
 
     # Type promotion to RealExpr works when we add a real-valued expr
-    children = [to_real(a), to_real(b[1]), RealExpr(:const, AbstractExpr[], 2.0, "const_2.0")]
-    @test isequal(sum([a, 1.0, 1, false, b[1]]), RealExpr(:add, children, nothing, Satisfiability.__get_hash_name(:add, children, is_commutative=true)))
+    children = [to_real(α), to_real(b[1]), RealExpr(:const, AbstractExpr[], 2.0, "const_2.0")]
+    @test isequal(sum([α, 1.0, 1, false, b[1]]), RealExpr(:add, children, nothing, Satisfiability.__get_hash_name(:add, children, is_commutative=true)))
 
     # Sum works automatically
-    @test isequal(1 + div(a, b[1]) + mod(b[1], b[2]) + true, sum([1, div(a, b[1]), mod(b[1], b[2]), true]))
+    @test isequal(1 + div(α, b[1]) + mod(b[1], b[2]) + true, sum([1, div(α, b[1]), mod(b[1], b[2]), true]))
 
[…]
