+++ test/solver_interface_tests.jl
@@ -5,28 +5,28 @@ using Test, Logging
 # assign is used after calling the solver so it belongs here.
 @testset "Assign values" begin
     @satvariable(x[1:3], Bool)
-    @satvariable(y[1:2], Bool)
+    @satvariable(β[1:2], Bool)
     @satvariable(z, Bool)
     
     prob = and(
         and(x),
-        and(x .∨ [y; z]),
-        and(¬y),
+        and(x .∨ [β; z]),
+        and(¬β),
         z
     )
     values = Dict{String, Bool}("x_1" => 1,"x_2" => 1,"x_3" => 1,
-              "y_1" => 0, "y_2" => 0,)
+              "$(Satisfiability.convert_to_ascii("β"))_2" => 0, "$(Satisfiability.convert_to_ascii("β"))_2" => 0,)
               assign!(prob, values)
     @test ismissing(value(z))
     z.value = 0
 
     @test all(value(x) .== [1, 1 ,1])
-    @test all(value(y) .== [0, 0])
+    @test all(value(β) .== [0, 0])
 
     # Creating a new expression where all children have assigned values also yields assigned values
-    @test all(value(x .∨ [y; z]) .== 1) 
-    @test all(value(xor.(x, [y; z])) .== 1) 
-    @test all(value(x .∧ [y; z]) .== 0) 
+    @test all(value(x .∨ [β; z]) .== 1) 
+    @test all(value(xor.(x, [β; z])) .== 1) 
+    @test all(value(x .∧ [β; z]) .== 0) 
     @test value(and(prob.children[1], prob.children[2])) == 1
 
     
