### 2024-06-16 elsoroka/Satisfiability.jl#49 @elsoroka
Sorry for the delay, busy week.

## Paper
> 2. I found it strange to have the "The SMT-LIB specification language" featured so prominently in the paper, only to conclude by stating that knowledge of SMT-LIB is not needed to use the package. It might be worth demoting this section to another part of the paper, in order to focus on your own work
This seems like a reasonable change.

## Repository
The CONTRIBUTING.md file and issue templates can be added - great idea! I'll take care of that in the next couple days.

## Code
Yes, the warnings are expected. Hopefully there is some way to suppress them during testing.
Thanks for your suggestion on running the examples as part of the tests. I agree this would prevent them from becoming out of date - same with the README. Does anyone know how to run code in a README automatically?

## Documentation
Yes! this is a recent change and now the documentation needs to get fixed. Thank you for catching it.
The `paper_examples` are from a previous submission. I agree the examples should be better organized and can get on that.
