### 2024-06-17 6cb6f61
Update issue templates
Per JOSS review, adding issue templates
