+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,15 @@
+---
+name: Feature request
+about: Suggest an idea for this project
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**If this request is about a new feature, please describe what you would like to see implemented.**
+
+**If this request is about an existing feature, what is the current behavior and what is the use case for changing the behavior?**
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
