### 2024-06-17 7ef42a2
refer to automatic installation of z3
