### 2024-06-18 #6757 @rafaelbailo
Hi @diehlpk, @elsoroka has began work addressing my comments in [the review issue](https://github.com/elsoroka/Satisfiability.jl/issues/49).
