### 2024-06-28 elsoroka/Satisfiability.jl#50 @diehlpk
Hi @elsoroka could you please have a look.
