### 2024-07-02 elsoroka/Satisfiability.jl#50 @elsoroka
Hi! yes, so sorry this has been delayed. I am working at an internship over the summer and have limited time to work on it -- we get a couple days off for the holiday this week so I'm targeting end of the week to address these comments.
