### 2024-07-07 929c42b
added to the README to more clearly indicate why there is a line calling Yices, due to issue #48.
