### 2024-07-07 4ab9ed8
Fix confusing mess about Z3 being installed automatically
