### 2024-07-08 elsoroka/Satisfiability.jl#49 @rafaelbailo
> Does anyone know how to run code in a README automatically?

I had a similar issue recently when developing [ConsensusBasedX.jl](https://github.com/PdIPS/ConsensusBasedX.jl). For the documentation examples, my solution was to keep the examples as full Julia script in a [separate folder](https://github.com/PdIPS/ConsensusBasedX.jl/tree/main/examples), to write a [unit test](https://github.com/PdIPS/ConsensusBasedX.jl/blob/main/test/examples.jl) to load and run each file and check that no errors occur, and then to modify the [documentation build](https://github.com/PdIPS/ConsensusBasedX.jl/blob/main/docs/make.jl) to inject each of these examples into the raw markdown files before the build process, using the custom syntax `{{subfolder/file}}` (here's [an example](https://github.com/PdIPS/ConsensusBasedX.jl/blob/main/docs/src/function_minimisation.md)). There are other ways of achieving the same (e.g. Literate.jl or Quarto), but I thought they were overkill for my particular need.

This unfortunately does not work for the `README.md` file. I have just submitted a [small pull request](https://github.com/elsoroka/Satisfiability.jl/pull/54) that tests the code in the `README.md` file. The exact output is not tested, only that the `julia` blocks can be parsed as Julia code and run without error. Given that these are only simple examples, and the methods should already be tested elsewhere, I think that's sufficient quality control.
