### 2024-07-08 b1447e3
Install Yices before CI
