+++ .github/workflows/ci.yml
@@ -14,21 +14,24 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@master
-      
+
       - name: Setup z3
         run: sudo apt-get -y install z3
-      
+
+      - name: Setup Yices
+        run: sudo add-apt-repository ppa:sri-csl/formal-methods; sudo apt-get update; sudo apt-get install yices2
+
       - name: Set up Julia latest version
         uses: julia-actions/setup-julia@latest
       - uses: julia-actions/cache@v2
-      
+
       - name: Run tests
         uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
       - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v4 
+        uses: codecov/codecov-action@v4
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         with:
-          file: lcov.info  # https://github.com/julia-actions/julia-processcoverage
+          file: lcov.info # https://github.com/julia-actions/julia-processcoverage
