### 2024-07-13 elsoroka/Satisfiability.jl#49 @elsoroka
Hi Rafael,

Yes, thank you so much for resolving this.
I've been struggling to respond in a timely manner because of my internship...I just can't scrape up any time to resolve these on weekdays.
