### 2024-07-13 6decd1e
Clean up warnings and spurious output in tests.
