### 2024-07-13 9998463
Clear out old paper_examples, which weren't interesting (leftover from the VMCAI submission) and add doctests to all example files. Doctests pass.
