+++ docs/make.jl
@@ -10,7 +10,6 @@ makedocs(sitename="Satisfiability.jl",
 #repo = Documenter.Remotes.GitHub("elsoroka", "Satisfiability.jl"),
 clean=true,
 draft=false,
-checkdocs=:none,
 source  = "src",
 workdir=pwd(),
 modules = [Satisfiability],
