### 2024-07-13 b37aa4e
moved small bitvector example to docs
