### 2024-07-13 elsoroka/Satisfiability.jl#50 @elsoroka
Addressing comments:

-   [x] The tests run, but I encountered one error and several warnings based on the main branch I pulled tonight (commit [dd89b1b](https://github.com/elsoroka/Satisfiability.jl/commit/dd89b1b8cd6b1d2f51cc3d67e96c12e14422ec06)). The tests that ran did pass, but I don't know if there were tests that didn't run due to errors, or if the warnings are something I should worry about.
PR #55 should suppress warnings in the tests.

- [x] I do not see a CONTRIBUTING.md for contributing to the repository. I also don't see a CODE_OF_CONDUCT.md or issue templates. Please consider adding these.
