### 2024-07-14 bfc4e6a
Added code coverage to some warnings and errors
