### 2024-07-14 c73483a
Added doctest for pages, but it doesn't work yet
