+++ docs/src/tutorial.md
@@ -7,32 +7,48 @@ We say a formula is **valid** if it is true for every assignment of values to it
 
 One famous transformation is De Morgan's law: `a ∧ b = ¬(¬a ∨ ¬b)`. To show validity of De Morgan's law, we can construct the bidirectional implication `a ∧ b ⟺ ¬(¬a ∨ ¬b)`. It suffices to show that the negation of this formula is unsatisfiable.
 
-```@example 1
+```jldoctest; output = false
 using Satisfiability 
 @satvariable(a, Bool)
 @satvariable(b, Bool)
 
 conjecture = iff(a ∧ b, ¬(¬a ∨ ¬b))
-status = sat!(¬conjecture, solver=Z3()) # status will be either :SAT or :UNSAT
+status = sat!(¬conjecture, solver=Z3()) # status should be :UNSAT
+
+# output
+
+UNSAT
 ```
 
-## A common logical mistake
-Suppose you have Boolean variables `p`, `q` and `r`. A common mistake made by students in discrete math classes is to think that if `p` implies `q` and `q` implies `r` (`(p ⟹ q) ∧ (q ⟹ r)`) then `p` must imply `r` (`p ⟹ r`). Are these statements equivalent? We can use a SAT solver to check.
+## Implications
+Suppose you have Boolean variables `p`, `q` and `r`. We want to find out: if `p` implies `q` and `q` implies `r` (`(p ⟹ q) ∧ (q ⟹ r)`) will `p` imply `r` (`p ⟹ r`)? We can use a SAT solver to check.
 
-```@example 1
+```jldoctest label9; output = false
+using Satisfiability
 @satvariable(p, Bool)
 @satvariable(q, Bool)
 @satvariable(r, Bool)
 
-conjecture = iff((p ⟹ q) ∧ (q ⟹ r), p ⟹ r)
+conjecture = ((p ⟹ q) ∧ (q ⟹ r)) ⟹ (p ⟹ r)
 status = sat!(¬conjecture, solver=Z3())
-```
-Unlike the previous example the status is `:SAT`, indicating there is an assignment `p`, `q` and `r` that disproves the conjecture.
 
-```@example 1
+# output
+
+UNSAT
+```
+Since the status is UNSAT, there is no example that disproves the conjecture. We see that the variables have taken on the value `nothing`.
+
+
+```jldoctest label9; output = false
 println("p = $(value(p))")
 println("q = $(value(q))")
 println("r = $(value(r))")
+
+# output
+
+p = nothing
+q = nothing
+r = nothing
 ```
 
 ## Optimizing over integers
@@ -41,21 +57,30 @@ The knapsack problem is a famous NP-complete problem in which you are packing a 
 A simpler version, illustrated in this [classic XKCD strip](https://xkcd.com/287/), is to pack the bag to exactly its maximum weight (or spend a specific amount of money).
 In fact, the problem in the XKCD strip can be expressed as a linear equation over integers.
 
-```@example 1
+```jldoctest; output = false 
+using Satisfiability
 @satvariable(a[1:6], Int)
 c = [215; 275; 335; 355; 420; 580]
+
 expr = and(and(a .>= 0), sum(a .* c) == 1505)
 sat!(expr, solver=Z3())
+
 println("Result: $(value(a))")
 println("Check: $(sum(value(a) .* c))")
+
+# output
+
+Result: [7, 0, 0, 0, 0, 0]
+Check: 1505
 ```
 
 ## Proving properties of fixed-size integer arithmetic
 This example is from Microsoft's [Z3 tutorial](https://microsoft.github.io/z3guide/docs/theories/Bitvectors/).
[…]
