### 2024-07-14 f714dd7
fixed doctests, they run again
