+++ docs/src/advanced.md
@@ -58,6 +58,7 @@ close(interactive_solver)
 # output
 
 status = unsat
+
 ()
 ```
 
