### 2024-07-14 edece9b
Don't fail on missing docstrings
