### 2024-07-14 d3a516f
propagate adding Yices
