+++ .github/workflows/docs.yml
@@ -16,6 +16,8 @@ jobs:
       steps:
         - name: Setup z3
           run: sudo apt-get -y install z3
+        - name: Setup Yices
+          run: sudo add-apt-repository ppa:sri-csl/formal-methods; sudo apt-get update; sudo apt-get install yices2
         - uses: actions/checkout@v4
         - uses: julia-actions/cache@v2
         - uses: julia-actions/julia-buildpkg@v1
