### 2024-07-14 elsoroka/Satisfiability.jl#49 @elsoroka
Hi @rafaelbailo and @diehlpk,
I've addressed the various documentation issues (reorganizing examples, cleaning them up etc) and code issues, including increasing unittests via `@test_throws` and fixing instances where the error or warning was printed while running the unittest.

I also made some edits to the paper to spend less time discussing the SMT-LIB language. A silly question: where/how should I commit those edits?
Thank you,
Emi
