### 2024-07-14 elsoroka/Satisfiability.jl#49 @rafaelbailo
Hi @elsoroka, thanks for that! I think the code and documentation changes should be committed to `main`, unless you have a specific need not to do so. Ultimately, the changes will have to be merged into `main` and a new release made before the paper is accepted. As for the paper changes, I would commit them to the `joss-paper` branch so that we can automatically compile the edited version of the manuscript.
