### 2024-07-14 e962f2f
Addressing reviewer comments: less emphasis on SMT-LIB and explicitly state that using the SMT-LIB language allows different solvers to be used.
