+++ joss-paper/paper.md
@@ -32,14 +32,12 @@ Theorem proving software is one of the core tools in formal verification, model 
 This paper introduces `Satisfiability.jl`, a package providing a high-level representation for SMT formulae including propositional logic, integer and real-valued arithmetic, and bit vectors in Julia [@bezanson2017julia]. `Satisfiability.jl` is the first published package for SMT solving in idiomatic Julia, taking advantage of language features such as multiple dispatch and metaprogramming to simplify the process of specifying and solving an SMT problem.
 
 ## The SMT-LIB specification language
-SMT-LIB is a low-level specification language designed to standardize interactions with theorem provers.
-The current SMT-LIB standard is V2.6; we used this version of the language specification when implementing our software. (To disambiguate between the SMT (satisfiability modulo theories) and this specification language, we always refer to the language as SMT-LIB.) There are both in-depth treatments of computational logic and associated decision procedures [@bradley2007calculus; @kroening2016decision] and shorter overviews of SMT [@de2011satisfiabilityintro].
+SMT-LIB is a low-level specification language designed to standardize interactions with theorem provers[^1].
+(To disambiguate between the SMT (satisfiability modulo theories) and this specification language, we always refer to the language as SMT-LIB.) There are both in-depth treatments of computational logic and associated decision procedures [@bradley2007calculus; @kroening2016decision] and shorter overviews of SMT [@de2011satisfiabilityintro].
 
-SMT-LIB uses a Lisp-like syntax designed to simplify input parsing, providing an interactive interface for theorem proving similar to Julia's REPL. SMT-LIB supports declaring variables, defining functions, making assertions (requiring that a Boolean predicate be true), and issuing solver commands. SMT expressions also have an associated _sort_, which constrains what functions or operations are valid for a given expression. For example, the SMT-LIB variable declaration `(declare-fun a () Int)` declares a symbol `a` with sort `Int`, and the command `(define-fun f () Bool (> a 1))` defines a function `a > 1` with sort `Bool`. The concept of SMT sorts maps cleanly onto Julia's type system.
+SMT-LIB uses a Lisp-like syntax designed to simplify input parsing, providing an interactive interface for theorem proving similar to Julia's REPL. The language supports declaring variables, defining functions, making assertions (requiring that a Boolean predicate be true), and issuing solver commands. However, a limitation of SMT-LIB is that many commands are only valid in specific solver modes. The command `(get-model)`, for example, retrieves the satisfying assignment for a predicate and is only valid in `sat` mode, while `(get-unsat-core)` is only valid in `unsat` mode. Issuing a command in the wrong mode yields an error, thus many useful sequences of SMT-LIB commands cannot be scripted in advance.
 
-A limitation of SMT-LIB is that many commands are only valid in specific solver modes. For example, the command `(get-model)` retrieves the satisfying assignment for a predicate and is only valid in `sat` mode, while `(get-unsat-core)` is only valid in `unsat` mode. Issuing a command in the wrong mode yields an error, thus many useful sequences of SMT-LIB commands cannot be scripted in advance.
-
-For a full description of SMT-LIB, readers are referred to the standard [@smtlib2]. As our software provides an abstraction on top of SMT-LIB, we refrain from an in-depth description of the language. Knowledge of SMT-LIB is not required to use `Satisfiability.jl`.
+For a full description of SMT-LIB, readers are referred to the standard [@smtlib2]. Because our software provides an abstraction on top of SMT-LIB (thus offering compatibility with different SMT solver backends), we refrain from an in-depth description of the language. Knowledge of SMT-LIB is not required to use `Satisfiability.jl`.
 
 # Statement of need
[…]
