### 2024-07-14 elsoroka/Satisfiability.jl#49 @elsoroka
OK @rafaelbailo I have committed the paper changes to `joss-paper`. The code and documentation changes have already been merged into `main`.
Does there need to be a new tagged software release? The latest tag was May 28 which was the latest change to software functionality. The documentation was automatically rebuilt with the new changes yesterday.
