### 2024-07-15 elsoroka/Satisfiability.jl#49 @rafaelbailo
There's no need for a new release until the editor instructs it. This will be right before publication. I'll go over your changes ASAP!
