### 2024-07-15 elsoroka/Satisfiability.jl#49 @rafaelbailo
Everything looks good, closing this as completed, I'll update the review issue as well. Thanks for all your work, and congratulations on a great project! 😄
