### 2024-07-17 #6757 @diehlpk
> Hi @diehlpk sorry for the confusion. I didn't understand where to post but I think all reviewer comments were just finished in these two threads
> https://github.com/elsoroka/Satisfiability.jl/issues/49
> https://github.com/elsoroka/Satisfiability.jl/issues/50

Yes, you can post here or on the tickets opened by the reviewers.
