### 2024-07-24 elsoroka/Satisfiability.jl#50 @computablee
@elsoroka I tried the following for CVC4 and similar for Alt-Ergo:
```julia
using Satisfiability

@satvariable(p, Bool)
@satvariable(q, Bool)
@satvariable(r, Bool)

conjecture = ((p ⟹ q) ∧ (q ⟹ r)) ⟹ (p ⟹ r)
status = sat!(!conjecture, solver=CVC4())
println(status)
```

Receiving:

```
ERROR: LoadError: UndefVarError: `CVC4` not defined
Stacktrace:
 [1] top-level scope
   @ ~/prop.jl:8
```

As for everything else, the only thing remaining that I haven't seen is a clear list of dependencies. Once that is resolved and I hear a response from this comment, I think I'll be satisfied with the submission. Almost everything is checked on my checklist.
