### 2024-07-26 #6757 @diehlpk
> Apologies for the delay in my review. Here are some observations.
>
> This is a _very_ nice package, providing key functionality for the Julia ecosystem. The documentation is very well done, with nice motivating examples.
>
>     * The package looks for a command-line z3 function. This should ideally be provided instead via BinaryBuilder. (E.g. [Z3.jl](https://github.com/ahumenberger/Z3.jl) does this.)
>
>     * Does the command-line interface work on Windows?
>
>     * On MacOS `brew install z3` should work (assuming the user has Homebrew installed) to install Z3.
>
>     * Why is the function called `sat!` instead of `sat`? It's not clear what it's mutating.
>
>     * I suggest to add a simple usage example to the paper. This would help focus the reader's understanding of the functionality and API of the package.
>
>     * There doesn’t seem to be any API documentation.
>
>     * There don’t seem to be any contribution guidelines.
>
>
> Overall this is a very nice piece of work!

@elsoroka could you please have a look at these comments?
