### 2024-07-26 #6757 @elsoroka
- [ ] The package looks for a command-line z3 function. This should ideally be provided instead via BinaryBuilder. (E.g. [Z3.jl](https://github.com/ahumenberger/Z3.jl) does this.)
So a recent update by Prof. Kochenderfer added using Z3.jl as you recommended, however other solvers that may not have a BinaryBuilder can be invoked by command-line. This is a good catch so when I have time tomorrow I will edit the paper to mention this.

- [ ] Does the command-line interface work on Windows?
At some point I verified this. I will put this on tomorrow's task list to check again including the use of BinaryBuilder for Z3.

- [x] On MacOS `brew install z3` should work (assuming the user has Homebrew installed) to install Z3.
Got it. I added this change to [the documentation](https://elsoroka.github.io/Satisfiability.jl/dev/installation/#Installing-Satisfiability).

- [ ] Why is the function called `sat!` instead of `sat`? It's not clear what it's mutating.
If a satisfying assignment exists, the assignment gets stored in the .value property of the variables/expressions, thus modifying them. I can add a sentence to the paper clarifying this - would it be helpful?

- [x] I suggest to add a simple usage example to the paper. This would help focus the reader's understanding of the functionality and API of the package.
In [the other reviewers' issue](https://github.com/elsoroka/Satisfiability.jl/issues/49#issuecomment-2155441360), it was suggested to avoid code examples in the paper because they could become out of date in future versions. Apologies for missing this in this thread earlier.

- [ ] There doesn’t seem to be any API documentation.
There is a page on the documentation called [Functions](https://elsoroka.github.io/Satisfiability.jl/dev/functions/) which lists all the exported functions. Does this count, or if not what changes would you recommend?

- [x] There don’t seem to be any contribution guidelines.
Also caught! [There are now.](https://elsoroka.github.io/Satisfiability.jl/dev/contributing/)
