### 2024-07-26 elsoroka/Satisfiability.jl#50 @elsoroka
**Not a fix yet:** I tested this for CVC4 and I can see two things:
There isn't a Solver object defined for CVC4 in Satisfiability.jl But we can make our own using a command like
`CVC4 =  Solver("CVC4", `cvc4 --lang smt2 --interactive --produce-models --qqq`)`
Then `sat!(expr, solver=CVC4)` should work, but doesn't return...I think something is wrong with how the output of CVC4 is being parsed by my code.
see: https://elsoroka.github.io/Satisfiability.jl/dev/advanced/#Custom-solver-options-and-using-other-solvers

I'm looking into why this is and will respond with an update.
