### 2024-07-28 elsoroka/Satisfiability.jl#50 @computablee
@elsoroka I'm not aware of such an option. There is a Python API and a C++ API for CVC4, but I realize implementing either of those is asking a lot when CVC5 is already supported and better in almost all ways in my experience. I know the Why3 platform can interface with both CVC4 and CVC5... it might be worth checking how they do it.

In my opinion, just knowing that CVC4 is not supported is sufficient. I wasn't aware of the `Solver()` call to load in a new SMT solver, so I'm fine on my end to mark that as solved.
