### 2024-07-28 elsoroka/Satisfiability.jl#50 @elsoroka
## Does Alt-Ergo work
I'm actually not sure how to test this. The reason is Satisfiability.jl relies on opening an interactive session, so we can do things like decide whether to say `(get-model)` after `(check-sat)` (because `(get-model)` isn't a valid command if `(check-sat)` is invalid).
I read the documentation and I am not sure whether this is a thing. I only see examples of file input. https://ocamlpro.github.io/alt-ergo/latest/Usage/index.html

This has overall been helpful because now we have more examples to discuss in the documentation. And I will make a Git issue about CVC4 and your suggestions for other packages that work with it @computablee.

## Dependencies
They are in the Project.toml file https://github.com/elsoroka/Satisfiability.jl/blob/main/Project.toml
