### 2024-07-28 deec625
Documentation updates to describe using a custom solver, per issue #50
