### 2024-08-03 elsoroka/Satisfiability.jl#50 @elsoroka
@computablee are you ready to mark this thread solved?
