### 2024-08-04 cbaf472
add support for 3d tensors
