### 2024-08-04 85e6c1b
add tests for 3d tensors
