### 2024-08-04 ba11aab
make tests compatible with windows
