### 2024-08-09 elsoroka/Satisfiability.jl#50 @diehlpk
@computablee could you please have a look?
