### 2024-08-09 elsoroka/Satisfiability.jl#50 @computablee
So sorry about the late response! Work has been busy. Yes, I think my issues have been solved. I'll mark it on the review thread. Would you like me to close this issue?
