### 2024-08-11 elsoroka/Satisfiability.jl#50 @elsoroka
I think since the review thread is solved @computablee, this one can be closed. Thank you!
