### 2024-08-12 elsoroka/Satisfiability.jl#50 @computablee
Sounds good! Closing @elsoroka
