### 2024-08-17 91ee264
Increment version number and add line in documentation for new release per JOSS publication checklist.
