+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Satisfiability"
 uuid = "160ab843-0bc6-4ba4-9585-b7478b70f443"
 authors = ["Emi Soroka <esoroka@stanford.edu>, Mykel Kochenderfer, Sanjay Lall"]
-version = "0.1.1"
+version = "0.1.3"
 
 [compat]
 julia = "1"
