### 2024-08-17 6b4f73f
i reread the version rules and think it should be called 0.2.0 because new features are added
