### 2024-08-17 #6757 @elsoroka
- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. Version is 0.2.0!
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here: 10.6084/m9.figshare.26768461

@diehlpk good news, I got these tasks done!
