### 2024-08-18 #6757 @diehlpk
@elsoroka the title of the archive should read as

Satisfiability.jl: Satisfiability Modulo Theories in Julia

and not as in the attached image:

![Screenshot_20240818_172132_DuckDuckGo.jpg](https://github.com/user-attachments/assets/3b60b30b-c58a-46bd-b7b5-f116971f72ac)
