+++ joss-paper/bibliography.bib
@@ -134,7 +134,7 @@
 
 @TECHREPORT{smtlib2,
 	author =	 {Clark Barrett and Pascal Fontaine and Cesare Tinelli},
-	title =	 {{The SMT-LIB Standard: Version 2.6}},
+	title =	 {The {SMT-LIB} Standard: Version 2.6},
 	institution =	 {Department of Computer Science, The University of Iowa},
 	year =	 2017,
 	note =	 {Available at {\tt www.SMT-LIB.org}}
@@ -173,7 +173,7 @@
 }
 
 @inproceedings{mann2021smt,
-	title={SMT-switch: a solver-agnostic C++ API for SMT solving},
+	title={SMT-switch: a solver-agnostic {C}++ API for SMT solving},
 	author={Mann, Makai and Wilson, Amalee and Zohar, Yoni and Stuntz, Lindsey and Irfan, Ahmed and Brown, Kristopher and Donovick, Caleb and Guman, Allison and Tinelli, Cesare and Barrett, Clark},
 	booktitle={International Conference on Theory and Applications of Satisfiability Testing},
 	pages={377--386},
