### 2024-08-19 #6757 @danielskatz
@elsoroka - I just found 2 small changes needed in the references, as indicated in https://github.com/elsoroka/Satisfiability.jl/pull/66

Please merge this, or let me now what you disagree with, then we can finish the publication process.
