### 2024-08-20 #6757 @danielskatz
👋 @elsoroka - the changes I made in the PR didn't make it in to this proof, I think because I was working in your main branch but this review thread is working in joss-paper. Should JOSS be working in main now? If so, we can change this thread to point to main.  If not, can you make the same two changes in the joss-paper branch?
