### 2024-08-20 #6757 @elsoroka
Hi @danielskatz  - to solve this, I merged the changes from main into joss-paper. So now the two branches should be in sync [https://github.com/elsoroka/Satisfiability.jl/compare/main...joss-paper](https://github.com/elsoroka/Satisfiability.jl/compare/main...joss-paper).
I think it's fine to work off either.
