### 2024-08-20 #6757 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
