### 2024-08-26 24339c5
adding citation and paper badge!
